# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Conference App
  version: '1.0.0'
servers:
  - url: "https://conference.nathakusuma.com/api/v1"
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRole:
      type: string
      enum: [ user, admin, event_coordinator ]
...
    UserMinimal:
      type: object
      properties:
        id:
          type: [ "string", "null" ]
          format: uuid
        name:
          type: [ "string", "null" ]
          examples:
            - "Natha Kusuma"
        role:
          oneOf:
            - $ref: '#/components/schemas/UserRole'
            - type: "null"
          examples:
            - "user"
        bio:
          type: [ "string", "null" ]
          examples:
            - "Seorang programmer pemula yang sedang belajar backend"
    ConferenceStatus:
      type: string
      enum: [ pending, approved, rejected ]
      examples:
        - "approved"
    Conference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          examples:
            - "Konferensi Programmer Backend"
        description:
          type: string
          examples:
            - "Membahas masalah backend terkini"
        speaker_name:
          type: string
          examples:
            - "Bambang Bimbang"
        speaker_title:
          type: string
          examples:
            - "Sepuh backend"
        target_audience:
          type: string
          examples:
            - "Junior Backend Developer di Malang Raya"
        prerequisites:
          type: [ string, "null" ]
          examples:
            - "Memiliki pemahaman dasar tentang bahasa pemrograman Go"
        seats:
          type: integer
          minimum: 1
          examples:
            - 10
        starts_at:
          type: string
          format: date-time
          examples:
            - "2025-01-28T01:04:40+07:00"
        ends_at:
          type: string
          format: date-time
          examples:
            - "2025-01-29T01:04:40+07:00"
        host:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
              examples:
                - "Natha Kusuma"
        status:
          $ref: '#/components/schemas/ConferenceStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        seats_taken:
          type: [ integer, "null" ]
          examples:
            - 0
...
    Pagination:
      type: object
      properties:
        has_more:
          type: boolean
          description: Indicates if there are more items available
          examples:
            - true
        first_id:
          type: string
          format: uuid
          description: ID of the first item in the current page
          examples:
            - "0194831b-9072-4090-803e-ffb74d52eb5c"
        last_id:
          type: string
          format: uuid
          description: ID of the last item in the current page
          examples:
            - "019497b5-0072-404b-804a-371d9a961ac6"
    ErrorResponse:
      type: object
      properties:
        message:
          type: [ "string", "null" ]
          description: Error message
        detail:
          type: [ "object", "null" ]
          description: Additional error details
        error_code:
          type: [ "string", "null" ]
          description: Error code identifier
        trace_id:
          type: [ "string", "null" ]
          format: uuid
          description: Unique trace ID for error tracking
  responses:
...
    AuthenticationError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noBearerToken:
              summary: No Bearer Token
              value:
                message: "You're not logged in. Please login first."
                error_code: "NO_BEARER_TOKEN"
            invalidBearerToken:
              summary: Invalid Bearer Token
              value:
                message: "Your auth session is invalid. Please renew your auth session."
                error_code: "INVALID_BEARER_TOKEN"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Data not found."
            error_code: "NOT_FOUND"
    FailParseRequest:
      description: Failed to parse request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Failed to parse request. Please check your request format."
            error_code: "FAIL_PARSE_REQUEST"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Something went wrong in our server. Please try again later."
            error_code: "INTERNAL_SERVER_ERROR"
            trace_id: "652e0a03-1c0a-404a-ac47-685d7ecb22d8"
    ForbiddenRole:
      description: Forbidden - User role not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "You're not allowed to access this resource."
            error_code: "FORBIDDEN_ROLE"
...
paths:
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get User by ID
      operationId: getUserById
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "01949e48-9f6b-796b-9611-3c9025493233"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: '#/components/schemas/UserMinimal'
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Users
      summary: Delete User
      description: Deletes a user. Only available to users with admin role.
      operationId: deleteUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "01949e48-9f6b-796b-9611-3c9025493233"
      responses:
        '204':
          description: Success - User deleted successfully
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/ForbiddenRole'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /conferences/{id}:
    get:
      tags:
        - Conferences
      summary: Get conference by ID
      description: Get a conference by its ID. Available to all roles.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Conference ID
      responses:
        '200':
          description: Successfully retrieved conference
          content:
            application/json:
              schema:
                type: object
                properties:
                  conference:
                    $ref: '#/components/schemas/Conference'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '403':
          description: Forbidden - Host is other user and is not approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "You're not allowed to access this resource."
                error_code: "FORBIDDEN_USER"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Conferences
      summary: Edit conference details
      description: Update an existing conference's details. Available to users with user role. This method allows partial updates.
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Conference ID
          example: "01948d68-4872-4007-8096-517b5e9bf6c2"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: [ string, "null" ]
                  minLength: 3
                  maxLength: 100
                  examples:
                    - "Konferensi Programmer Backend Part 3"
                description:
                  type: [ string, "null" ]
                  minLength: 3
                  maxLength: 1000
                  examples:
                    - "Membahas teknologi backend terkini"
                speaker_name:
                  type: [ string, "null" ]
                  minLength: 3
                  maxLength: 100
                  examples:
                    - "Budi Idbu"
                speaker_title:
                  type: [ string, "null" ]
                  minLength: 3
                  maxLength: 100
                  examples:
                    - "The real sepuh of para sepuh"
                target_audience:
                  type: [ string, "null" ]
                  minLength: 3
                  maxLength: 255
                  examples:
                    - "Senior Backend Developer di Malang"
                prerequisites:
                  type: [ string, "null" ]
                  maxLength: 255
                  examples:
                    - "Memiliki pemahaman mendalam tentang bahasa pemrograman Go"
                seats:
                  type: [ integer, "null" ]
                  minimum: 1
                  examples:
                    - 10
                starts_at:
                  type: [ string, "null" ]
                  format: date-time
                  examples:
                    - "2025-02-01T10:50:00.000000Z"
                ends_at:
                  type: [ string, "null" ]
                  format: date-time
                  examples:
                    - "2025-02-01T14:45:00.000000Z"
      responses:
        '204':
          description: Conference successfully updated
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          description: Forbidden - User not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbiddenUser:
                  summary: Forbidden User
                  value:
                    message: "You're not allowed to access this resource."
                    error_code: "FORBIDDEN_USER"
                forbiddenRole:
                  summary: Forbidden Role
                  value:
                    message: "You're not allowed to access this resource."
                    error_code: "FORBIDDEN_ROLE"
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Time window conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "There's already a conference in the same time window. Please choose another time window."
                detail:
                  conferences:
                    - id: "0194831b-9072-4090-803e-ffb74d52eb5c"
                      title: "Description for future conference 1"
                      starts_at: "2025-02-01T09:46:49.330992Z"
                      ends_at: "2025-02-01T11:46:49.330992Z"
                error_code: "TIME_WINDOW_CONFLICT"
        '422':
          description: Validation or business rule error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                updateApprovedTimeWindow:
                  summary: Update Approved Time Window
                  value:
                    message: "You're not allowed to update an approved conference time window."
                    error_code: "UPDATE_APPROVED_TIME_WINDOW"
                updateNotPendingConference:
                  summary: Update Not Pending Conference
                  value:
                    message: "You're not allowed to update a conference that is not pending"
                    error_code: "UPDATE_NOT_PENDING_CONFERENCE"
                timeAlreadyPassed:
                  summary: Time Already Passed
                  value:
                    message: "Time has already passed. Please use future time."
                    error_code: "TIME_ALREADY_PASSED"
                endTimeBeforeStart:
                  summary: End Time Before Start
                  value:
                    message: "End time is before start time. Please use correct time."
                    error_code: "END_TIME_BEFORE_START"
                updatePastConference:
                  summary: Update Past Conference
                  value:
                    message: "You're not allowed to update a past conference."
                    error_code: "UPDATE_PAST_CONFERENCE"
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Conferences
      summary: Delete conference
      description: Delete an existing conference. Available to users with user and event_coordinator roles.
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Conference ID
          example: "01948d68-4872-4007-8096-517b5e9bf6c2"
      responses:
        '204':
          description: Conference successfully deleted
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          description: Forbidden - User not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbiddenUser:
                  summary: Not Author or Event Coordinator
                  value:
                    message: "You're not allowed to access this resource."
                    error_code: "FORBIDDEN_USER"
                forbiddenRole:
                  summary: Forbidden Role
                  value:
                    message: "You're not allowed to access this resource."
                    error_code: "FORBIDDEN_ROLE"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /registrations:
    post:
      tags:
        - Registrations
      summary: Register for a conference
      description: Register for a conference. Available to users with user role.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - conference_id
              properties:
                conference_id:
                  type: string
                  format: uuid
                  example: "0194890f-f8d4-406c-804d-1c007a44e144"
      responses:
        '200':
          description: Successfully registered for conference
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notRoleUser:
                  summary: Not Role User
                  value:
                    message: "You're not allowed to access this resource."
                    error_code: "FORBIDDEN_ROLE"
                hostRegisterOwn:
                  summary: Host Register Own Conference
                  value:
                    message: "You're not allowed to register to your own conference."
                    error_code: "HOST_CANNOT_REGISTER"
                conferenceNotApproved:
                  summary: Conference Not Approved
                  value:
                    message: "You're not allowed to access this resource."
                    error_code: "FORBIDDEN_USER"
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                alreadyRegistered:
                  summary: Already Registered
                  value:
                    message: "You're already registered to this conference."
                    error_code: "USER_ALREADY_REGISTERED_TO_CONFERENCE"
                fullSeat:
                  summary: Full Seat
                  value:
                    message: "Conference is full. You're not allowed to register anymore."
                    error_code: "CONFERENCE_FULL"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conferenceEnded:
                  summary: Conference Ended
                  value:
                    message: "Conference has ended. You're not allowed to register anymore."
                    error_code: "CONFERENCE_ENDED"
                validation:
                  summary: Validation Error
                  value:
                    message: "There are invalid fields in your request. Please check and try again"
                    detail:
                      - conference_id:
                          tag: "required"
                          param: ""
                          translation: "ConferenceID is a required field"
                    error_code: "VALIDATION_ERROR"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /registrations/conferences/{id}:
    get:
      tags:
        - Registrations
      summary: Get registered users for a conference
      description: Get registered users for a conference. Available to all roles. For users with user role, only their hosted conferences are allowed.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "01945590-60d4-4062-80a2-2effbfa127fe"
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 20
          example: 10
        - name: after_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: before_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved registered users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                users:
                  - id: "019412d4-a354-4068-809e-6545dcf6aca7"
                    name: "User 2"
                  - id: "0194130b-91d4-4095-80e2-5af8f28930e0"
                    name: "User 3"
                pagination:
                  has_more: false
                  first_id: "019412d4-a354-4068-809e-6545dcf6aca7"
                  last_id: "0194130b-91d4-4095-80e2-5af8f28930e0"
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation:
                  summary: Validation Error
                  value:
                    message: "There are invalid fields in your request. Please check and try again"
                    detail:
                      - limit:
                          tag: "required"
                          param: ""
                          translation: "Limit is a required field"
                    error_code: "VALIDATION_ERROR"
                invalidPagination:
                  summary: Invalid Pagination
                  value:
                    message: "Cannot use after_id and before_id at the same time."
                    error_code: "INVALID_PAGINATION"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /registrations/users/{id}:
    get:
      tags:
        - Registrations
      summary: Get registered conferences for a user
      description: Get registered conferences for a user. Available to all roles. For users with user role, only their own id is allowed.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "019412d4-a354-4068-809e-6545dcf6aca9"
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 20
          example: 10
        - name: include_past
          in: query
          required: false
          schema:
            type: boolean
          example: true
        - name: after_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: before_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved registered conferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  conferences:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conference'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/FailParseRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "You're not allowed to access this resource."
                error_code: "FORBIDDEN_USER"
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation:
                  summary: Validation Error
                  value:
                    message: "There are invalid fields in your request. Please check and try again"
                    detail:
                      - limit:
                          tag: "required"
                          param: ""
                          translation: "Limit is a required field"
                    error_code: "VALIDATION_ERROR"
                invalidPagination:
                  summary: Invalid Pagination
                  value:
                    message: "Cannot use after_id and before_id at the same time."
                    error_code: "INVALID_PAGINATION"
        '500':
          $ref: '#/components/responses/InternalServerError'
...